*: fix test
